refactor: split classes from OkioChannels.kt into different files
