+++ io-lib/src/commonMain/kotlin/it/unibo/tuprolog/solve/libs/io/channel/SourceInputChannel.kt
@@ -1,8 +1,6 @@
 package it.unibo.tuprolog.solve.libs.io.channel
 
 import it.unibo.tuprolog.solve.channel.impl.AbstractInputChannel
-import it.unibo.tuprolog.solve.channel.impl.AbstractOutputChannel
-import okio.BufferedSink
 import okio.BufferedSource
 
 /**
@@ -32,26 +30,6 @@ internal class SourceInputChannel(
 
     override fun close() {
         source.close()
-        super.close()
-    }
-}
-
-/**
- * An [it.unibo.tuprolog.solve.channel.OutputChannel] writing UTF-8 text to an Okio [BufferedSink].
- */
-internal class SinkOutputChannel(
-    private val sink: BufferedSink,
-) : AbstractOutputChannel<String>() {
-    override fun writeActually(value: String) {
-        sink.writeUtf8(value)
-    }
-
-    override fun flushActually() {
-        sink.flush()
-    }
-
-    override fun close() {
-        sink.close()
         super.close()
     }
 }
